fix: Fix e_commerce product list paging with invalid `start` query parameter
To handle the pagination in the product list on the website,
a `start` query parameter is added to the URL. This parameter
is read both by the JS client code and by the back-end.

When this `start` parameter was invalid (i.e. a negative number
or a non-number string), the JS did crash (because of a JSON.parse),
and the backend retrieved bad data.

This has been fixed by clamping the `start` parameter to a positive
or zero value on the back-end, and on the front-end, both when emitting
the value (setting the query parameter), and when reading the value.

The `data-start` HTML attribute of the button used to show the previous
page has also been clamped to a positive/zero value, which is not
strictly needed because of the other changes.

Note: no special handling of the next page button is done. This means
that, for suffiently large values of `start`, no product is shown in
the product list, and no button is shown to go to another page.
This behavior is not related to this commit.